i5-3789 - Traverse Elasticsearch query strings for date keywords and replace them.
