+++ lib/index.js
@@ -73,7 +73,7 @@ exports.jsonPatchSchema = require('./json-patch-schema');
 
 exports.Aggregation = require('./aggregation').Aggregation;
 exports.AggregateExpression = require('./aggregate-expression').AggregateExpression;
-exports.elasticDateTranslator = require('./elastic-query-traverser');
+exports.esQueryTransformer = require('./elastic-query-transformer');
 exports.DateKeywords = require('./date-keywords').DateKeywords;
 
 exports.exists = function (value, exception) {
